% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{my_function}
\alias{my_function}
\title{My important function}
\usage{
my_function()
}
\value{
This fun thing
}
\description{
My important function
}
\section{This function is similar to \code{base}{
:split()}:

Here is why
}

